+++ test/commands/source/deploy.test.ts
@@ -12,11 +12,11 @@ import sinon from 'sinon';
 import { expect } from 'chai';
 import { ComponentSetBuilder, ComponentSetOptions, MetadataApiDeployOptions } from '@salesforce/source-deploy-retrieve';
 import { fromStub, stubInterface, stubMethod } from '@salesforce/ts-sinon';
+import { AnyJson } from '@salesforce/ts-types';
 import { ConfigAggregator, Lifecycle, Messages, SfProject } from '@salesforce/core';
 import { Config } from '@oclif/core';
 import { SfCommand } from '@salesforce/sf-plugins-core';
 import { MockTestOrgData, TestContext } from '@salesforce/core/lib/testSetup.js';
-import ConfigMeta, { ConfigVars } from '@salesforce/plugin-deploy-retrieve/lib/configMeta.js';
 import { Deploy } from '../../../src/commands/force/source/deploy.js';
 import { DeployCommandResult, DeployResultFormatter } from '../../../src/formatters/deployResultFormatter.js';
 import {
@@ -87,8 +87,26 @@ describe('force:source:deploy', () => {
   const runDeployCmd = async (params: string[], options?: { sourceApiVersion?: string }) => {
     const cmd = new TestDeploy(params, oclifConfigStub);
     cmd.project = SfProject.getInstance();
-
-    cmd.configAggregator = await (await ConfigAggregator.create({ customConfigMeta: ConfigMeta.default })).reload();
+    cmd.configAggregator =
+      await // this is simplified from https://github.com/salesforcecli/plugin-deploy-retrieve/blob/d481ac1d79e29e20302924e688be7ec75df22957/src/configMeta.ts#L24-L25
+      // as a workaround for
+      // 1. PDR being ESM before plugin-source
+      // 2. lib/configMeta.js not being in the top-level exports
+      (
+        await ConfigAggregator.create({
+          customConfigMeta: [
+            {
+              key: 'org-metadata-rest-deploy',
+              description: 'foo',
+              hidden: true,
+              input: {
+                validator: (value: AnyJson): boolean => typeof value === 'string' && ['true', 'false'].includes(value),
+                failedMessage: 'boo!',
+              },
+            },
+          ],
+        })
+      ).reload();
     sandbox.stub(cmd.project, 'getDefaultPackage').returns({ name: '', path: '', fullPath: defaultDir });
     sandbox.stub(cmd.project, 'getUniquePackageDirectories').returns([{ fullPath: defaultDir, path: '', name: '' }]);
     sandbox.stub(cmd.project, 'getPackageDirectories').returns([{ fullPath: defaultDir, path: '', name: '' }]);
@@ -418,7 +436,7 @@ describe('force:source:deploy', () => {
       });
 
       it('should REST deploy with config', async () => {
-        await $$.stubConfig({ [ConfigVars.ORG_METADATA_REST_DEPLOY]: 'true', 'target-org': testOrg.username });
+        await $$.stubConfig({ 'org-metadata-rest-deploy': 'true', 'target-org': testOrg.username });
 
         const manifest = 'package.xml';
         const result = await runDeployCmd(['--manifest', manifest, '--json']);